don't change classes but add attributes instead
this is a fundamental change to the pipeline. formerly, as we went through successive functions, new classes were added to our objects. with the update to R 4.0 (and tidyverse update) managing classes is a lot more complicated. instead, we will be storing info on the steps we've taken in the attributes() of each object. hoping this will be more robust in the long term...