make read_motive_csv functional
OK I think the function below mostly works. I haven't really tested it thoroughly, but in principle it should be able to flexibly handle Motive CSVs that have rigid bodies and/or markers
(including files that mix the two). I've written some comments on checks that could be inserted to ensure things are working.