change variables to singular case
used the feature in notepad++ to Find in Files and then replace all instances of _lengths, _widths, and _heights to _length, etc. these replacements were done on the .R files in the R folder and the example_scripts folders only. i then used Install and Restart in rstudio to regenerate the package. everything seems to have worked out fine...i think...