add barebones first drafts of roxygen
these are really simple drafts -- i am putting no effort in to detail for any of them. rather, the point is so that they get added to NAMESPACE and get "exported" to the enviroment. this should make these functions accessible when using library(pathviewR). note that i have not done this for all functions; just the ones i have been using & editing